cleaned up app module
